+++ model/des/write_res0_des.f90
@@ -23,7 +23,7 @@ MODULE WRITE_RES0_DES_MODULE
 
       DOUBLE PRECISION, DIMENSION(:), INTENT(IN) :: des_radius, ro_sol
       DOUBLE PRECISION, DIMENSION(:,:), INTENT(IN) :: des_vel_new, des_pos_new, omega_new, des_usr_var
-      INTEGER(KIND=1), DIMENSION(:), INTENT(IN) :: particle_state
+      INTEGER, DIMENSION(:), INTENT(IN) :: particle_state
       INTEGER, DIMENSION(:), INTENT(IN) :: iglobal_id
 
 !-----------------------------------------------
